Rebrand the completion code so far as the completion SchemaManager
